Refactor Settings Menu/Command Palette entries
This commit makes use of ST3's `edit_settings` command.
