+++ Context.sublime-menu
@@ -3,13 +3,13 @@
         "caption": "-"
     },
     {
+        "caption": "Navigate to Definition",
         "command": "navigate_to_definition",
-        "args": {},
-        "caption": "Navigate to Definition"
+        "args": {}
     },
     {
+        "caption": "Jump Back",
         "command": "jump_prev",
-        "args": {},
-        "caption": "Jump Back"
+        "args": {}
     }
 ]
