+++ plugins/utils.py
@@ -13,8 +13,8 @@ try:
     from sublime import status_message, error_message
 
 except ImportError:  # running tests
-    from ..tests.sublime_fake import sublime
-    from ..tests.sublime_fake import sublime_plugin
+    from .tests.sublime_fake import sublime
+    from .tests.sublime_fake import sublime_plugin
 
     sys.modules['sublime'] = sublime
     sys.modules['sublime_plugin'] = sublime_plugin
